+++ apps/tanstack-web/app/root-provider.tsx
@@ -1,47 +1,6 @@
-import { QueryClient } from "@tanstack/react-query";
-import superjson from "superjson";
-import { createTRPCClient, httpBatchStreamLink } from "@trpc/client";
-import { createTRPCOptionsProxy } from "@trpc/tanstack-react-query";
-import { TRPCProvider } from "./trpc/react";
-import { TrpcRouter } from "@habinook/trpc";
+import { queryClient, TRPCProvider, trpcClient } from "./trpc";
 
-const queryClient = new QueryClient({
-	defaultOptions: {
-		dehydrate: { serializeData: superjson.serialize },
-		hydrate: { deserializeData: superjson.deserialize },
-	},
-});
-
-function getUrl() {
-	const base = (() => {
-		if (typeof window !== "undefined") return "";
-		return `http://localhost:${process.env.PORT ?? 3000}`;
-	})();
-	return base + "/api/trpc";
-}
-
-export const trpcClient = createTRPCClient<TrpcRouter>({
-	links: [
-		httpBatchStreamLink({
-			transformer: superjson,
-			url: getUrl(),
-		}),
-	],
-});
-
-const serverHelpers = createTRPCOptionsProxy({
-	client: trpcClient,
-	queryClient: queryClient,
-});
-
-export function getContext() {
-	return {
-		queryClient,
-		trpc: serverHelpers,
-	};
-}
-
-export function Provider({ children }: { children: React.ReactNode }) {
+export function TrpcProvider({ children }: { children: React.ReactNode }) {
 	return (
 		<TRPCProvider trpcClient={trpcClient} queryClient={queryClient}>
 			{children}
