--- apps/tanstack-web/app/trpc/react.ts
@@ -1,4 +1,0 @@
-import { TrpcRouter } from "@habinook/trpc";
-import { createTRPCContext } from "@trpc/tanstack-react-query";
-
-export const { TRPCProvider, useTRPC } = createTRPCContext<TrpcRouter>();
